backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
PyMuPDF==1.24.4
numpy==1.26.4
scikit-learn==1.5.0
google-cloud-storage==2.17.0
pinecone-client==4.1.0
google-generativeai==0.7.2
httpx==0.27.0
weasyprint==61.2
Jinja2==3.1.4
pydantic==2.8.2
prometheus-client==0.20.0
python-jose==3.3.0
psycopg[binary]==3.2.1
supabase==2.6.0
reportlab==4.2.2

